+++ common/transaction.go
@@ -1,4 +1,4 @@
-package transaction
+package common
 
 // Transaction interface는 it-chain에서 사용되는 모든 Transaction이 공통적으로 가져야 하는 함수들을 정의함.
 type Transaction interface {
